de-dupe filters, and update copy
