+++ static/gsApp/views/seerAutomation/scm.tsx
@@ -1,35 +1,18 @@
-import {parseAsString, parseAsStringLiteral, useQueryState} from 'nuqs';
-
-import {CompactSelect, type SelectOption} from '@sentry/scraps/compactSelect';
-import {Input} from '@sentry/scraps/input';
 import {Flex} from '@sentry/scraps/layout';
 import {ExternalLink} from '@sentry/scraps/link';
 
 import AnalyticsArea from 'sentry/components/analyticsArea';
 import {ScmIntegrationTree} from 'sentry/components/repositories/scmIntegrationTree/scmIntegrationTree';
-import type {RepoFilter} from 'sentry/components/repositories/scmIntegrationTree/types';
+import ScmTreeFilters from 'sentry/components/repositories/scmIntegrationTree/scmTreeFilters';
+import useScmTreeFilters from 'sentry/components/repositories/scmIntegrationTree/useScmTreeFilters';
 import {t, tct} from 'sentry/locale';
 import SettingsPageHeader from 'sentry/views/settings/components/settingsPageHeader';
 
 import SeerSettingsPageContent from 'getsentry/views/seerAutomation/components/seerSettingsPageContent';
 import SeerSettingsPageWrapper from 'getsentry/views/seerAutomation/components/seerSettingsPageWrapper';
 
-const REPO_FILTER_OPTIONS: Array<SelectOption<RepoFilter>> = [
-  {value: 'all' as const, label: t('All repos')},
-  {value: 'connected' as const, label: t('Connected Repos')},
-  {value: 'not-connected' as const, label: t('Disconnected Repos')},
-];
-
-const repoParser = parseAsStringLiteral(
-  REPO_FILTER_OPTIONS.map(option => option.value)
-).withDefault('all');
-
 export default function SeerAutomationSCM() {
-  const [searchTerm, setSearchTerm] = useQueryState(
-    'search',
-    parseAsString.withDefault('')
-  );
-  const [repoFilter, setRepoFilter] = useQueryState('repo', repoParser);
+  const {repoFilter, setRepoFilter, searchTerm, setSearchTerm} = useScmTreeFilters();
 
   return (
     <AnalyticsArea name="scm">
@@ -37,7 +20,7 @@ export default function SeerAutomationSCM() {
         <SettingsPageHeader
           title={t('Seer SCM Config')}
           subtitle={tct(
-            `Install an SCM Integration to connect your source code to Seer. Seer needs read access to your source code to perform code review, and analyze your issues. [read_the_docs:Read the docs] and our [privacy:AI Privacy Principles] to learn more.`,
+            `Integrate with a Seer compatible Source Code Management provider and then connect repositories with Sentry. Seer needs read access to your source code to perform code review, and analyze your issues. [read_the_docs:Read the docs] and our [privacy:AI Privacy Principles] to learn more.`,
             {
               privacy: (
                 <ExternalLink href="https://docs.sentry.io/product/ai-in-sentry/ai-privacy-and-security/" />
@@ -50,19 +33,13 @@ export default function SeerAutomationSCM() {
         />
         <SeerSettingsPageContent>
           <Flex align="center" gap="md">
-            <CompactSelect
-              value={repoFilter}
-              onChange={(opt: SelectOption<RepoFilter>) => setRepoFilter(opt.value)}
-              options={REPO_FILTER_OPTIONS}
-            />
-            <Input
-              value={searchTerm}
-              onChange={e => setSearchTerm(e.target.value)}
-              placeholder={t('Search repos\u2026')}
-              style={{flex: 1}}
+            <ScmTreeFilters
+              repoFilter={repoFilter}
+              setRepoFilter={setRepoFilter}
+              searchTerm={searchTerm}
+              setSearchTerm={setSearchTerm}
             />
           </Flex>
-
           <ScmIntegrationTree
             providerFilter="seer-supported"
             repoFilter={repoFilter}
